Refactor: Migrasikan sisa panel admin ke xoradmin
Melanjutkan refactoring besar pada panel admin dengan memigrasikan fungsionalitas yang tersisa ke dalam panel `/xoradmin` terpusat.

**Perubahan Utama**:
- Fungsionalitas dari `UserController`, `BalanceController`, `LogController`, `ForwardManagerController`, dan `DebugFeedController` telah dimigrasikan sepenuhnya ke dalam `XorAdminController`.
- Halaman-halaman ini sekarang dapat diakses sebagai tab baru di dalam antarmuka `/xoradmin`.
- Semua controller dan view yang lama dan berlebihan untuk fitur-fitur di atas telah dihapus dari codebase.
- Entri `CHANGELOG.md` telah diperbarui untuk mencerminkan progres migrasi ini.
