+++ src/Views/admin/xoradmin/index.php
@@ -1,5 +1,13 @@
 <?php
 // This view assumes all necessary data is available in the $data array.
+if (!function_exists('format_currency')) {
+    function format_currency($number, $currency = 'Rp') {
+        if (is_null($number)) {
+            $number = 0;
+        }
+        return $currency . ' ' . number_format($number, 0, ',', '.');
+    }
+}
 ?>
 <!DOCTYPE html>
 <html lang="id">
@@ -49,6 +57,9 @@
 
         <div class="tabs">
             <a href="?action=dashboard" class="tab-link <?= $data['active_tab'] === 'dashboard' ? 'active' : '' ?>">Dashboard</a>
+            <a href="?action=users" class="tab-link <?= $data['active_tab'] === 'users' ? 'active' : '' ?>">Manajemen Pengguna</a>
+            <a href="?action=balance" class="tab-link <?= $data['active_tab'] === 'balance' ? 'active' : '' ?>">Manajemen Saldo</a>
+            <a href="?action=logs" class="tab-link <?= $data['active_tab'] === 'logs' ? 'active' : '' ?>">Logs</a>
             <a href="?action=bots" class="tab-link <?= $data['active_tab'] === 'bots' || $data['active_tab'] === 'edit_bot' ? 'active' : '' ?>">Manajemen Bot</a>
             <a href="?action=roles" class="tab-link <?= $data['active_tab'] === 'roles' ? 'active' : '' ?>">Manajemen Peran</a>
             <a href="?action=db_reset" class="tab-link <?= $data['active_tab'] === 'db_reset' ? 'active' : '' ?>">Reset Database</a>
@@ -136,6 +147,137 @@
                 <?php endif; ?>
             </div>
 
+            <?php elseif ($data['active_tab'] === 'users'): ?>
+                <div id="users" class="tab-content active">
+                    <h2>Manajemen Pengguna</h2>
+                    <div class="search-form" style="margin-bottom: 20px;">
+                        <form action="?" method="get">
+                            <input type="hidden" name="action" value="users">
+                            <input type="text" name="search" placeholder="Cari ID, Nama, Username..." value="<?= htmlspecialchars($data['search_term']) ?>" style="width: 300px; display: inline-block;">
+                            <button type="submit" style="width: auto; display: inline-block;">Cari</button>
+                            <?php if(!empty($data['search_term'])): ?>
+                                <a href="?action=users">Hapus Filter</a>
+                            <?php endif; ?>
+                        </form>
+                    </div>
+
+                    <p>Menampilkan <?= count($data['users']) ?> dari <?= $data['total_users'] ?> total pengguna.</p>
+
+                    <table>
+                        <thead>
+                            <tr>
+                                <th>User ID</th>
+                                <th>Nama</th>
+                                <th>Username</th>
+                                <th>Status</th>
+                                <th>Peran</th>
+                                <th>Aksi</th>
+                            </tr>
+                        </thead>
+                        <tbody>
+                            <?php if (empty($data['users'])): ?>
+                                <tr><td colspan="6" style="text-align:center;">Tidak ada pengguna ditemukan.</td></tr>
+                            <?php else: ?>
+                                <?php foreach ($data['users'] as $user): ?>
+                                <tr>
+                                    <td><?= htmlspecialchars($user['id']) ?></td>
+                                    <td><?= htmlspecialchars(trim(($user['first_name'] ?? '') . ' ' . ($user['last_name'] ?? ''))) ?></td>
+                                    <td>@<?= htmlspecialchars($user['username'] ?? 'N/A') ?></td>
+                                    <td><?= htmlspecialchars(ucfirst($user['status'])) ?></td>
+                                    <td><?= htmlspecialchars($user['roles'] ?? 'Tidak ada') ?></td>
+                                    <td><a href="?action=dashboard&search_user=<?= htmlspecialchars($user['username'] ?? $user['id']) ?>">Lihat Chat</a></td>
+                                </tr>
+                                <?php endforeach; ?>
+                            <?php endif; ?>
+                        </tbody>
+                    </table>
+                    <div class="pagination" style="margin-top: 20px;">
+                        <?php
+                        $currentPage = $data['page'];
+                        $totalPages = $data['total_pages'];
+                        $query_params = ['action' => 'users', 'search' => $data['search_term']];
+
+                        if ($totalPages > 1):
+                            if ($currentPage > 1) {
+                                $query_params['page'] = $currentPage - 1;
+                                echo '<a href="?' . http_build_query($query_params) . '">&laquo; Sebelumnya</a>';
+                            }
+                            echo '<span style="margin: 0 10px;"> Halaman ' . $currentPage . ' dari ' . $totalPages . ' </span>';
+                            if ($currentPage < $totalPages) {
+                                $query_params['page'] = $currentPage + 1;
+                                echo '<a href="?' . http_build_query($query_params) . '">Berikutnya &raquo;</a>';
+                            }
+                        endif;
+                        ?>
+                    </div>
+                </div>
+            <?php elseif ($data['active_tab'] === 'balance'): ?>
+                <div id="balance" class="tab-content active">
+                    <h2>Manajemen Saldo</h2>
+                    <?php if (!empty($_SESSION['flash_message'])): ?>
+                        <div class="message"><?= htmlspecialchars($_SESSION['flash_message']); unset($_SESSION['flash_message']); ?></div>
+                    <?php endif; ?>
+
+                    <div class="search-form" style="margin-bottom: 20px;">
+                        <form action="?" method="get">
+                            <input type="hidden" name="action" value="balance">
+                            <input type="text" name="search" placeholder="Cari Nama/Username..." value="<?= htmlspecialchars($data['search_term']) ?>" style="width: 300px; display: inline-block;">
+                            <button type="submit" style="width: auto; display: inline-block;">Cari</button>
+                            <?php if(!empty($data['search_term'])): ?>
+                                <a href="?action=balance">Hapus Filter</a>
+                            <?php endif; ?>
+                        </form>
+                    </div>
+
+                     <table>
+                        <thead>
+                            <tr>
+                                <th>ID</th>
+                                <th>Nama</th>
+                                <th>Saldo</th>
+                                <th>Pemasukan</th>
+                                <th>Pengeluaran</th>
+                                <th>Aksi</th>
+                            </tr>
+                        </thead>
+                        <tbody>
+                            <?php foreach ($data['users_data'] as $user_data): ?>
+                                <tr>
+                                    <td><?= htmlspecialchars($user_data['telegram_id']) ?></td>
+                                    <td><?= htmlspecialchars(trim($user_data['first_name'] . ' ' . $user_data['last_name'])) ?></td>
+                                    <td class="clickable-log" data-log-type="balance" data-telegram-id="<?= $user_data['telegram_id'] ?>" data-user-name="<?= htmlspecialchars($user_data['first_name']) ?>"><?= format_currency($user_data['balance']) ?></td>
+                                    <td class="clickable-log" data-log-type="sales" data-telegram-id="<?= $user_data['telegram_id'] ?>" data-user-name="<?= htmlspecialchars($user_data['first_name']) ?>"><?= format_currency($user_data['total_income']) ?></td>
+                                    <td class="clickable-log" data-log-type="purchases" data-telegram-id="<?= $user_data['telegram_id'] ?>" data-user-name="<?= htmlspecialchars($user_data['first_name']) ?>"><?= format_currency($user_data['total_spending']) ?></td>
+                                    <td>
+                                        <button class="open-balance-modal" data-telegram-id="<?= $user_data['telegram_id'] ?>" data-user-name="<?= htmlspecialchars(trim($user_data['first_name'] . ' ' . $user_data['last_name'])) ?>">
+                                            Ubah Saldo
+                                        </button>
+                                    </td>
+                                </tr>
+                            <?php endforeach; ?>
+                        </tbody>
+                    </table>
+
+                    <div class="pagination" style="margin-top: 20px;">
+                         <?php
+                        $currentPage = $data['page'];
+                        $totalPages = $data['total_pages'];
+                        $query_params = ['action' => 'balance', 'search' => $data['search_term']];
+
+                        if ($totalPages > 1):
+                            if ($currentPage > 1) {
+                                $query_params['page'] = $currentPage - 1;
+                                echo '<a href="?' . http_build_query($query_params) . '">&laquo; Sebelumnya</a>';
+                            }
+                            echo '<span style="margin: 0 10px;"> Halaman ' . $currentPage . ' dari ' . $totalPages . ' </span>';
+                            if ($currentPage < $totalPages) {
+                                $query_params['page'] = $currentPage + 1;
+                                echo '<a href="?' . http_build_query($query_params) . '">Berikutnya &raquo;</a>';
+                            }
+                        endif;
+                        ?>
+                    </div>
+                </div>
             <?php elseif ($data['active_tab'] === 'bots'): ?>
                 <div id="bots" class="tab-content active">
                     <h2>Manajemen Bot</h2>
@@ -277,6 +419,120 @@
                         <?php endif; ?>
                     <?php endif; ?>
                 </div>
+            <?php elseif ($data['active_tab'] === 'logs'): ?>
+                <div id="logs" class="tab-content active">
+                    <h2>Logs</h2>
+                    <div class="tabs">
+                        <a href="?action=logs&type=app" class="tab-link <?= $data['log_type'] === 'app' ? 'active' : '' ?>">App Logs</a>
+                        <a href="?action=logs&type=media" class="tab-link <?= $data['log_type'] === 'media' ? 'active' : '' ?>">Media Logs</a>
+                        <a href="?action=logs&type=telegram" class="tab-link <?= $data['log_type'] === 'telegram' ? 'active' : '' ?>">Telegram Error Logs</a>
+                        <a href="?action=debug_feed" class="tab-link <?= $data['active_tab'] === 'debug_feed' ? 'active' : '' ?>">Debug Feed</a>
+                    </div>
+                    <div class="log-content" style="padding-top: 20px;">
+                        <?php if ($data['log_type'] === 'app'): ?>
+                            <div>
+                                <form action="?" method="get" style="display:inline-block;">
+                                    <input type="hidden" name="action" value="logs">
+                                    <input type="hidden" name="type" value="app">
+                                    <label for="level_select">Filter by Level:</label>
+                                    <select name="level" id="level_select" onchange="this.form.submit()">
+                                        <option value="all" <?= ($data['selected_level'] === 'all') ? 'selected' : '' ?>>All Levels</option>
+                                        <?php foreach ($data['log_levels'] as $level): ?>
+                                            <option value="<?= htmlspecialchars($level) ?>" <?= ($data['selected_level'] === $level) ? 'selected' : '' ?>>
+                                                <?= ucfirst(htmlspecialchars($level)) ?>
+                                            </option>
+                                        <?php endforeach; ?>
+                                    </select>
+                                </form>
+                                <form action="/xoradmin/clearAppLogs" method="post" style="display:inline-block;" onsubmit="return confirm('Anda yakin ingin MENGHAPUS SEMUA log aplikasi?');">
+                                    <button type="submit">Clear App Logs</button>
+                                </form>
+                            </div>
+                            <table>
+                                <thead><tr><th>ID</th><th>Timestamp</th><th>Level</th><th>Message</th><th>Context</th></tr></thead>
+                                <tbody>
+                                    <?php foreach ($data['logs'] as $log): ?>
+                                    <tr>
+                                        <td><?= $log['id'] ?></td>
+                                        <td><?= $log['created_at'] ?></td>
+                                        <td><?= $log['level'] ?></td>
+                                        <td><?= htmlspecialchars($log['message']) ?></td>
+                                        <td>
+                                            <?php if (!empty($log['context'])): ?>
+                                                <button class="context-toggle-button" data-target="context-<?= $log['id'] ?>">Show</button>
+                                                <pre id="context-<?= $log['id'] ?>" style="display: none;"><code><?= htmlspecialchars(json_encode(json_decode($log['context']), JSON_PRETTY_PRINT))) ?></code></pre>
+                                            <?php endif; ?>
+                                        </td>
+                                    </tr>
+                                    <?php endforeach; ?>
+                                </tbody>
+                            </table>
+                            <!-- Pagination for App Logs here -->
+                        <?php elseif ($data['log_type'] === 'media'): ?>
+                            <table>
+                                <thead><tr><th>Waktu</th><th>Pengguna</th><th>Bot</th><th>Detail Media</th><th>Aksi</th></tr></thead>
+                                <tbody>
+                                    <?php foreach ($data['grouped_logs'] as $group): ?>
+                                    <tr>
+                                        <td><?= htmlspecialchars($group['group_info']['time']) ?></td>
+                                        <td><?= htmlspecialchars($group['group_info']['user']) ?></td>
+                                        <td><?= htmlspecialchars($group['group_info']['bot']) ?></td>
+                                        <td>
+                                            <?php foreach ($group['items'] as $item): ?>
+                                            <div><?= htmlspecialchars($item['type']) ?>: <?= htmlspecialchars($item['file_name'] ?? '') ?></div>
+                                            <?php endforeach; ?>
+                                        </td>
+                                        <td>
+                                            <button class="btn-forward" data-group-id="<?= htmlspecialchars($group['group_info']['media_group_id'] ?? 'single_' . $group['items'][0]['id']) ?>" data-bot-id="<?= htmlspecialchars($group['group_info']['bot_id']) ?>">
+                                                Teruskan
+                                            </button>
+                                        </td>
+                                    </tr>
+                                    <?php endforeach; ?>
+                                </tbody>
+                            </table>
+                        <?php elseif ($data['log_type'] === 'telegram'): ?>
+                             <table>
+                                <thead><tr><th>Waktu</th><th>Metode</th><th>Chat ID</th><th>Kode</th><th>Deskripsi</th><th>Request</th></tr></thead>
+                                <tbody>
+                                    <?php foreach ($data['logs'] as $log): ?>
+                                    <tr>
+                                        <td><?= htmlspecialchars($log['created_at']) ?></td>
+                                        <td><?= htmlspecialchars($log['method']) ?></td>
+                                        <td><?= htmlspecialchars($log['chat_id'] ?? 'N/A') ?></td>
+                                        <td><?= htmlspecialchars($log['error_code'] ?? 'N/A') ?></td>
+                                        <td><?= htmlspecialchars($log['description']) ?></td>
+                                        <td>
+                                            <?php if (!empty($log['request_data'])): ?>
+                                                <button class="context-toggle-button" data-target="context-<?= $log['id'] ?>">Show</button>
+                                                <pre id="context-<?= $log['id'] ?>" style="display: none;"><code><?= htmlspecialchars(json_encode(json_decode($log['request_data']), JSON_PRETTY_PRINT))) ?></code></pre>
+                                            <?php endif; ?>
+                                        </td>
+                                    </tr>
+                                    <?php endforeach; ?>
+                                </tbody>
+                            </table>
+                            <!-- Pagination for Telegram logs here -->
+                        <?php endif; ?>
+                    </div>
+                </div>
+            <?php elseif ($data['active_tab'] === 'debug_feed'): ?>
+                <div id="debug_feed" class="tab-content active">
+                    <h2>Raw Telegram Update Feed</h2>
+                    <table>
+                        <thead><tr><th>ID</th><th>Timestamp</th><th>Payload</th></tr></thead>
+                        <tbody>
+                            <?php foreach ($data['updates'] as $update): ?>
+                            <tr>
+                                <td><?= htmlspecialchars($update['id']) ?></td>
+                                <td><?= htmlspecialchars($update['created_at']) ?></td>
+                                <td><pre><code><?= htmlspecialchars(json_encode(json_decode($update['payload']), JSON_PRETTY_PRINT)) ?></code></pre></td>
+                            </tr>
+                            <?php endforeach; ?>
+                        </tbody>
+                    </table>
+                    <!-- Pagination -->
+                </div>
             <?php elseif ($data['active_tab'] === 'db_reset'): ?>
                 <div id="db_reset" class="tab-content danger-zone active">
                     <h2>Reset Database & Skema</h2>
@@ -294,17 +550,72 @@
 
     <div id="responseModal" class="modal"><div class="modal-content"><span class="close">&times;</span><h2 id="modal-title">Hasil Aksi</h2><pre id="modal-body">Memproses...</pre></div></div>
 
+    <!-- Modal untuk Ubah Saldo -->
+    <div id="balance-modal" class="modal" style="z-index: 1060;">
+        <div class="modal-content">
+            <div class="modal-header">
+                <h2 id="modal-title-adjust">Ubah Saldo untuk Pengguna</h2>
+                <span class="close-button close-balance-modal">&times;</span>
+            </div>
+            <form action="/xoradmin/adjustBalance?<?= http_build_query($_GET) ?>" method="post">
+                <input type="hidden" name="user_id" id="modal-user-id">
+                <label for="modal-amount">Jumlah:</label>
+                <input type="number" name="amount" id="modal-amount" step="0.01" min="0.01" required>
+                <label for="modal-description">Deskripsi:</label>
+                <input type="text" name="description" id="modal-description" required>
+                <button type="submit" name="action" value="add_balance">Tambah Saldo</button>
+                <button type="submit" name="action" value="subtract_balance">Kurangi Saldo</button>
+            </form>
+        </div>
+    </div>
+
+    <!-- Modal untuk Log Transaksi -->
+    <div id="log-modal" class="modal" style="z-index: 1070;">
+        <div class="modal-content" style="max-width: 800px;">
+            <div class="modal-header">
+                <h2 id="log-modal-title">Riwayat Transaksi</h2>
+                <span class="close-button close-log-modal">&times;</span>
+            </div>
+            <div id="log-modal-body" style="max-height: 60vh; overflow-y: auto;"></div>
+        </div>
+    </div>
+
     <script>
+        document.addEventListener('DOMContentLoaded', function() {
+            const toggleButtons = document.querySelectorAll('.context-toggle-button');
+            toggleButtons.forEach(button => {
+                button.addEventListener('click', function() {
+                    const targetId = this.getAttribute('data-target');
+                    const targetPre = document.getElementById(targetId);
+                    if (targetPre) {
+                        const isHidden = targetPre.style.display === 'none';
+                        targetPre.style.display = isHidden ? 'block' : 'none';
+                        this.textContent = isHidden ? 'Hide' : 'Show';
+                    }
+                });
+            });
+
+            document.querySelectorAll('.btn-forward').forEach(button => {
+                button.addEventListener('click', function() {
+                    const groupId = this.dataset.groupId;
+                    const botId = this.dataset.botId;
+                    if (confirm('Anda yakin ingin meneruskan media ini ke semua admin?')) {
+                        handleAjaxAction('forward_media', botId, { group_id: groupId });
+                    }
+                });
+            });
+        });
+
         const modal = document.getElementById('responseModal');
         const modalTitle = document.getElementById('modal-title');
         const modalBody = document.getElementById('modal-body');
         const span = document.getElementsByClassName('close')[0];
 
         function showModal(title, content) { modalTitle.textContent = title; modalBody.textContent = content; modal.style.display = 'block'; }
-        span.onclick = function() { modal.style.display = 'none'; }
+        if(span) span.onclick = function() { modal.style.display = 'none'; }
         window.onclick = function(event) { if (event.target == modal) modal.style.display = 'none'; }
 
-        async function handleAjaxAction(action, botId) {
+        async function handleAjaxAction(action, botId, extraData = {}) {
             let confirmation = action !== 'delete-webhook' || confirm('Yakin ingin menghapus webhook?');
             if (!confirmation) return;
 
@@ -313,6 +624,9 @@
                 const formData = new FormData();
                 formData.append('action', action);
                 if (botId) formData.append('bot_id', botId);
+                for(const key in extraData) {
+                    formData.append(key, extraData[key]);
+                }
 
                 const response = await fetch('/api/xoradmin', { method: 'POST', body: formData });
                 if (!response.ok) throw new Error(`HTTP error! status: ${response.status}`);
@@ -352,6 +666,87 @@
                 }
             });
         });
+
+        // Balance Modals Logic
+        const balanceModal = document.getElementById('balance-modal');
+        if(balanceModal) {
+            const balanceModalTitle = document.getElementById('modal-title-adjust');
+            const balanceModalUserIdInput = document.getElementById('modal-user-id');
+            document.querySelectorAll('.open-balance-modal').forEach(button => {
+                button.addEventListener('click', function() {
+                    balanceModalTitle.textContent = 'Ubah Saldo untuk ' + this.dataset.userName;
+                    balanceModalUserIdInput.value = this.dataset.telegramId;
+                    balanceModal.style.display = 'flex';
+                });
+            });
+            document.querySelectorAll('.close-balance-modal').forEach(button => button.addEventListener('click', () => balanceModal.style.display = 'none'));
+            if(balanceModal) balanceModal.addEventListener('click', e => { if (e.target === balanceModal) balanceModal.style.display = 'none'; });
+        }
+
+        const logModal = document.getElementById('log-modal');
+        if(logModal) {
+            const logModalTitle = document.getElementById('log-modal-title');
+            const logModalBody = document.getElementById('log-modal-body');
+            document.querySelectorAll('.clickable-log').forEach(cell => {
+                cell.addEventListener('click', async function() {
+                    const telegramId = this.dataset.telegramId;
+                    const logType = this.dataset.logType;
+
+                    logModalBody.innerHTML = '<p>Memuat data...</p>';
+                    logModal.style.display = 'flex';
+                    logModalTitle.textContent = `Riwayat untuk ${this.dataset.userName}`;
+
+                    try {
+                        const formData = new FormData();
+                        formData.append('action', `get_${logType}_log`);
+                        formData.append('telegram_id', telegramId);
+
+                        const response = await fetch('/api/xoradmin', { method: 'POST', body: formData });
+                        const result = await response.json();
+
+                        if (result.status !== 'success' || !result.data) throw new Error(result.message || 'Data tidak ditemukan.');
+
+                        if (result.data.length === 0) {
+                            logModalBody.innerHTML = '<p>Tidak ada riwayat ditemukan.</p>';
+                            return;
+                        }
+
+                        const table = document.createElement('table');
+                        table.className = 'chat-log-table';
+                        const thead = document.createElement('thead');
+                        const tbody = document.createElement('tbody');
+                        const headers = Object.keys(result.data[0]);
+
+                        let headerRow = '<tr>';
+                        headers.forEach(h => headerRow += `<th>${h}</th>`);
+                        headerRow += '</tr>';
+                        thead.innerHTML = headerRow;
+
+                        let tableBody = '';
+                        result.data.forEach(item => {
+                            tableBody += '<tr>';
+                            headers.forEach(h => {
+                                let val = item[h] || '';
+                                if(h.includes('price') || h.includes('amount')) val = new Intl.NumberFormat('id-ID', { style: 'currency', currency: 'IDR' }).format(val);
+                                tableBody += `<td>${val}</td>`
+                            });
+                            tableBody += '</tr>';
+                        });
+                        tbody.innerHTML = tableBody;
+
+                        table.appendChild(thead);
+                        table.appendChild(tbody);
+                        logModalBody.innerHTML = '';
+                        logModalBody.appendChild(table);
+
+                    } catch (error) {
+                        logModalBody.innerHTML = `<p style="color:red;">Gagal memuat data: ${error.message}</p>`;
+                    }
+                });
+            });
+            document.querySelectorAll('.close-log-modal').forEach(button => button.addEventListener('click', () => logModal.style.display = 'none'));
+            if(logModal) logModal.addEventListener('click', e => { if (e.target === logModal) logModal.style.display = 'none'; });
+        }
     </script>
 </body>
 </html>
